+++ internal/io/embedio/embedio.go
@@ -88,6 +88,7 @@ func (e *embedio) Query(emb []float32) (answer.Answer, error) {
 	var data []answer.Result
 	for i := range chs {
 		l := len(chs[i].PageIDs)
+		txt := e.txt.ChunkText(chs[i])
 		d := answer.Result{
 			ItemID:      chs[i].ItemID,
 			PageIDStart: chs[i].PageIDs[0],
@@ -98,6 +99,7 @@ func (e *embedio) Query(emb []float32) (answer.Answer, error) {
 				"https://www.biodiversitylibrary.org/page/",
 				chs[i].PageIDs[0],
 			),
+			Text: txt,
 		}
 		data = append(data, d)
 	}