+++ pkg/ent/text/interface.go
@@ -2,4 +2,5 @@ package text
 
 type Text interface {
 	TextToChunks(titleID uint) ([]Chunk, error)
+	ChunkText(Chunk) string
 }